cbor/requirements.txt
jaydebeapi==1.3.1
jpype1==1.5.0
